requirements.txt
streamlit>=1.24.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
google-generativeai==0.8.5
python-dotenv>=1.0.0
PyPDF2>=3.0.0
numpy>=1.24.0
tqdm>=4.65.0